requirements.txt
#

# These requirements were autogenerated by pipenv

# To regenerate from the project's Pipfile, run:

#

#    pipenv lock --requirements

#



-i https://pypi.org/simple

apscheduler==3.9.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4.0'

authlib==1.0.0

bcrypt==3.2.0

blinker==1.4

cachelib==0.6.0; python_version >= '3.6'

certifi==2021.10.8

cffi==1.15.0

charset-normalizer==2.0.12; python_version >= '3'

click==8.1.2; python_version >= '3.7'

colorama==0.4.4; platform_system == 'Windows'

cryptography==36.0.2

dnspython==2.2.1

flask-apscheduler==1.12.3

flask-caching==1.10.1

flask-login==0.6.0

flask-mail==0.9.1

flask-paranoid==0.2.0

flask-pymongo==2.3.0

flask-session==0.4.0

flask-sqlalchemy==2.5.1

flask==2.1.1

greenlet==1.1.2; python_version >= '3' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))

idna==3.3; python_version >= '3'

importlib-metadata==4.11.3; python_version < '3.10'

itsdangerous==2.1.2; python_version >= '3.7'

jinja2==3.1.1; python_version >= '3.7'

markupsafe==2.1.1; python_version >= '3.7'

psycopg2-binary==2.9.3

# psycopg2==2.9.3

pycparser==2.21; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

pycryptodome==3.14.1

pymongo==4.0.2; python_version >= '3.6'

python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

python-dotenv==0.20.0

pytz==2022.1

requests==2.27.1

six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

sqlalchemy-utils==0.38.2

sqlalchemy==1.4.34

tzlocal==2.1

urllib3==1.22

werkzeug==2.0.3

zipp==3.7.0; python_version >= '3.7'

